+++ Car.java
@@ -1,0 +1,32 @@
+import java.awt.*;
+
+abstract class Car extends Engine{
+    public int nrDoors; // Number of doors on the car
+    public double currentSpeed; // The current speed of the car
+    public Color color; // Color of the car
+    public String modelName;
+
+    public int getNrDoors(){
+        return nrDoors;
+    }
+
+    public double getCurrentSpeed(){
+        return currentSpeed;
+    }
+
+    public Color getColor(){
+        return color;
+    }
+
+    public void setColor(Color clr) {
+        color = clr;
+    }
+
+    public void startEngine(){
+        currentSpeed = 0.1;
+    }
+    public void stopEngine(){
+        currentSpeed = 0;
+    }
+
+}
