+++ Engine.java
@@ -1,0 +1,9 @@
+abstract class Engine {
+    public double enginePower; // Engine power of the engine
+
+    public double getEnginePower(){
+        return enginePower;
+    }
+    public void startEngine(){}
+    public void stopEngine(){}
+}
